backend/requirements.txt
fastapi
uvicorn
tensorflow-cpu
opencv-python-headless
python-multipart
numpy
pandas
scikit-learn
matplotlib
seaborn
groq>=0.4.0
python-dotenv
